emby-dev-test
testing emby implementation, possible avenue for jellyfin as well
