+++ auth-portal-full-stack-dev.env
@@ -1,6 +1,7 @@
 POSTGRES_PASSWORD=change-me-long-random
 SESSION_SECRET=change-me-32+chars-random
 APP_BASE_URL=http://localhost:8089
+MEDIA_PROVIDER=plex
 
 # Set 'FORCE_SECURE_COOKIE=1' in prod; Nginx should set 'X-Forwarded-Proto: https' anyway.
 FORCE_SECURE_COOKIE=0
@@ -11,8 +12,18 @@ DATA_KEY=5Z3UMPcF9BBkpB2SkuoXqYfGWKn1eXzpMdR8EyMV8dY=
 # LDAP (if using the ldap profile)
 LDAP_ADMIN_PASSWORD=change-me-strong
 
+# PLEX Variables
 # Authz (optional but recommended)
 PLEX_OWNER_TOKEN=plxxxxxxxxxxxxxxxxxxxx
 # Either set machine id or a server name (machine id wins if both present)
 PLEX_SERVER_MACHINE_ID=
 PLEX_SERVER_NAME=
+
+# EMBY Variables
+EMBY_SERVER_URL=http://localhost:8096
+EMBY_APP_NAME=AuthPortal
+EMBY_APP_VERSION=1.0.0
+# EMBY_QUICKCONNECT=1 (not yet ready)
+EMBY_API_KEY=
+EMBY_OWNER_USERNAME=
+EMBY_OWNER_ID
